+++ Lotus/src/utils/lotusError.cpp
@@ -16,8 +16,8 @@ void lotus::throwOverloadError(const String& overload, const String& type1, cons
     throw LotusException(type1 + STRING_LITERAL(": ") + STRING_LITERAL("No overload for ") + overload + STRING_LITERAL(" with ") + type2);
 }
 
-void lotus::checkThrowIndexError(const Value& index, int size, Module& module) {
-    if (index->asInt(module) < 0 || index->asInt(module) >= size) {
+void lotus::checkThrowIndexError(const Value& index, size_t size, Module& module) {
+    if (index->asInt(module) < 0 || static_cast<size_t>(index->asInt(module)) >= size) {
         THROW(module, STRING(INT(index->asInt(module))->asString(module) + STRING_LITERAL(" is invalid index")));
     }
 }
